finalize patches folders reorg
