+++ SolastaCommunityExpansion/Patches/CustomFeatures/RulesetCharacterHeroPatcher.cs
@@ -3,11 +3,39 @@ using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using HarmonyLib;
 using SolastaCommunityExpansion.CustomFeatureDefinitions;
+using SolastaCommunityExpansion.Helpers;
 using SolastaCommunityExpansion.Models;
 using SolastaModApi.Infrastructure;
 
 namespace SolastaCommunityExpansion.Patches.CustomFeatures
 {
+    [HarmonyPatch(typeof(RulesetCharacterHero), "EnumerateUsableRitualSpells")]
+    [SuppressMessage("Minor Code Smell", "S101:Types should be named in PascalCase", Justification = "Patch")]
+    internal static class RestModuleHitDice_EnumerateUsableRitualSpells_Patch
+    {
+        internal static void Postfix(RulesetCharacterHero __instance, RuleDefinitions.RitualCasting ritualType, List<SpellDefinition> ritualSpells)
+        {
+            if ((ExtraRitualCasting)ritualType != ExtraRitualCasting.Known) { return; }
+
+            var spellRepertoire = __instance.SpellRepertoires
+                .Where(r => r.SpellCastingFeature.SpellReadyness == RuleDefinitions.SpellReadyness.AllKnown)
+                .FirstOrDefault(r => r.SpellCastingFeature.SpellKnowledge == RuleDefinitions.SpellKnowledge.Selection);
+
+            if (spellRepertoire == null) { return; }
+
+            ritualSpells.AddRange(spellRepertoire.KnownSpells
+                .Where(s => s.Ritual)
+                .Where(s => spellRepertoire.MaxSpellLevelOfSpellCastingLevel >= s.SpellLevel));
+
+            if (spellRepertoire.AutoPreparedSpells == null) { return; }
+
+            ritualSpells.AddRange(spellRepertoire.AutoPreparedSpells
+                .Where(s => s.Ritual)
+                .Where(s => spellRepertoire.MaxSpellLevelOfSpellCastingLevel >= s.SpellLevel));
+
+        }
+    }
+
     [HarmonyPatch(typeof(RulesetCharacterHero), "FindClassHoldingFeature")]
     [SuppressMessage("Minor Code Smell", "S101:Types should be named in PascalCase", Justification = "Patch")]
     internal static class RulesetCharacterHero_FindClassHoldingFeature
