+++ SolastaCommunityExpansion/Patches/GameUi/AdventureLog/GameLocationBanterManagerPatcher.cs
@@ -2,7 +2,7 @@
 using System.Diagnostics.CodeAnalysis;
 using UnityEngine.AddressableAssets;
 
-namespace SolastaCommunityExpansion.Patches.AdventureLog
+namespace SolastaCommunityExpansion.Patches.GameUi.AdventureLog
 {
     [HarmonyPatch(typeof(GameLocationBanterManager), "ForceBanterLine")]
     [SuppressMessage("Minor Code Smell", "S101:Types should be named in PascalCase", Justification = "Patch")]
