+++ SolastaCommunityExpansion/Patches/SrdAndHouseRules/RulesetCharacterHeroPatcher.cs
@@ -1,9 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using HarmonyLib;
-using SolastaCommunityExpansion.Helpers;
 using static SolastaModApi.DatabaseHelper.QuestTreeDefinitions;
 
 namespace SolastaCommunityExpansion.Patches.SrdAndHouseRules
@@ -74,33 +72,6 @@ namespace SolastaCommunityExpansion.Patches.SrdAndHouseRules
                     Main.Log($"ComputeNeededExperienceToReachLevel: Dividing experience gained by {Main.Settings.MultiplyTheExperienceGainedBy}%. Original={original}, modified={__result}.");
                 }
             }
-        }
-    }
-
-    [HarmonyPatch(typeof(RulesetCharacterHero), "EnumerateUsableRitualSpells")]
-    [SuppressMessage("Minor Code Smell", "S101:Types should be named in PascalCase", Justification = "Patch")]
-    internal static class RestModuleHitDice_EnumerateUsableRitualSpells_Patch
-    {
-        internal static void Postfix(RulesetCharacterHero __instance, RuleDefinitions.RitualCasting ritualType, List<SpellDefinition> ritualSpells)
-        {
-            if ((ExtraRitualCasting)ritualType != ExtraRitualCasting.Known) { return; }
-
-            var spellRepertoire = __instance.SpellRepertoires
-                .Where(r => r.SpellCastingFeature.SpellReadyness == RuleDefinitions.SpellReadyness.AllKnown)
-                .FirstOrDefault(r => r.SpellCastingFeature.SpellKnowledge == RuleDefinitions.SpellKnowledge.Selection);
-
-            if (spellRepertoire == null) { return; }
-
-            ritualSpells.AddRange(spellRepertoire.KnownSpells
-                .Where(s => s.Ritual)
-                .Where(s => spellRepertoire.MaxSpellLevelOfSpellCastingLevel >= s.SpellLevel));
-
-            if (spellRepertoire.AutoPreparedSpells == null) { return; }
-
-            ritualSpells.AddRange(spellRepertoire.AutoPreparedSpells
-                .Where(s => s.Ritual)
-                .Where(s => spellRepertoire.MaxSpellLevelOfSpellCastingLevel >= s.SpellLevel));
-
         }
     }
 }
